Get rid of geometry hard coding (#109)
As reported in the follow-up to #99 and some other testing, parquet
files with names of columns that weren't 'geoemtry' weren't working.
This adds code to detect the geometry column name, and use it for
non-bbox requests and to sort the output.

Note that this will just use the first geometry column found - at some
point we should make this more robust and make use of the
`primary_column` in geoparquet, but I don't believe you can do that
directly with DuckDB (though I could be wrong...).